Automatically Merged using SST Master Branch Merger

+++ src/sst/elements/ember/shmem/motifs/emberShmemPut.h
@@ -117,7 +117,7 @@ public:
                 double latency = (totalTime/m_iterations);
                 printf("%d:%s: message-size %d, iterations %d, total-time %.3lf us, time-per %.3lf us, %.3f GB/s\n",m_my_pe,
                             getMotifName().c_str(),
-                            m_nelems * sizeof(TYPE),
+                            (int) (m_nelems * sizeof(TYPE)),
                             m_iterations,
                             totalTime * 1000000.0,
                             latency * 1000000.0,
